+++ example/minibase/slot_wood.py
@@ -17,6 +17,7 @@ ex_base = slot_wood(
     board_width_spacer = .2,
     board_break_width = .4,
     nail_diameter = .6,
+    nail_overlap_height = .4,
     joist_width = 4,
     debug = False
 )