+++ src/cqterrain/minibase/base_irregular.py
@@ -1,5 +1,20 @@
+# Copyright 2025 James Adams
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
 import cadquery as cq
-from . import circle, ellipse, hexagon, rectangle, slot
+from . import circle
 from cadqueryhelper import irregular_grid
 import random
 
@@ -45,7 +60,7 @@ def base_irregular(
         tile = tile_style(length, width, height)
         return tile
     
-    bounds = base.val().BoundingBox()
+    bounds = base.val().BoundingBox() #type:ignore
     xlen = bounds.xlen
     ylen = bounds.ylen
     zlen = bounds.zlen